Developpement

+++ src/app/components/category-list/category-list.component.ts
@@ -3,12 +3,14 @@ import { Component, inject, OnInit } from "@angular/core";
 
 import { CategoryService } from "../../services/category/category.service";
 import { CategoryFormComponent } from "../category-form/category-form.component";
+import { ModalDeleteComponent } from "../modal-delete/modal-delete.component";
+import { Delete_Modal_Type } from "../../types/Delete_Modal";
 
 @Component({
   selector: "app-category-list",
-  imports: [CategoryFormComponent],
+  imports: [CategoryFormComponent, ModalDeleteComponent],
   templateUrl: "./category-list.component.html",
-  styleUrl: "./category-list.component.css"
+  styleUrl: "./category-list.component.css",
 })
 export class CategoryListComponent implements OnInit {
   categoryList: CategoryRequest[] = [];
@@ -16,6 +18,14 @@ export class CategoryListComponent implements OnInit {
   titleModal?: string;
   isPopupVisible: boolean = false;
   categoryService = inject(CategoryService);
+  isModalDelete: boolean = false;
+  valueDeleteModal: Delete_Modal_Type = {
+    title: "Deleting category",
+    content: "Are you sure you want to delete the category ",
+    valueBtnAccepted: "Yes, delete the category",
+    valueBtnCancelled: "Cancel, keep the category",
+    entityToDelete: null,
+  };
 
   ngOnInit(): void {
     this.getCategory();
@@ -43,17 +53,31 @@ export class CategoryListComponent implements OnInit {
     if (isFetch) this.getCategory();
   }
 
-  deleteCategory(categoryRequest: CategoryRequest) {
-    if (categoryRequest.id) {
+  deleteCategory(categoryRequest: unknown) {
+    if (this.isCategoryRequest(categoryRequest) && categoryRequest.id) {
       this.categoryService.deleteCategory(categoryRequest.id).subscribe({
         complete: () => {
           this.getCategory();
-        }
+          this.closeModalDelete();
+        },
       });
     }
   }
 
+  openModalDelete(categoryRequest: CategoryRequest) {
+    this.valueDeleteModal.entityToDelete = categoryRequest;
+    this.isModalDelete = true;
+  }
+
+  closeModalDelete() {
+    this.isModalDelete = false;
+  }
+
   init() {
     this.category = { name: "" };
+  }
+
+  private isCategoryRequest(obj: unknown): obj is CategoryRequest {
+    return typeof obj === "object" && obj !== null && "name" in obj;
   }
 }
